% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypermean.R
\name{hypermean}
\alias{hypermean}
\title{hypermean}
\usage{
hypermean(x, size = 0.1, type = "trim", ret = "mean")
}
\arguments{
\item{x}{vector of observations}

\item{size}{trimming size. The value should be from 0 to 1.  (e.g. 0.1 is trimming 10\%)}

\item{type}{Type of weighting. Default is \code{"trim"}. \code{c("trim","wins","epa","triw","trig","gauss")}}

\item{ret}{Return value. Default is \code{"mean"}. \code{c("mean","vector")}}
}
\description{
This function computes hypermeans, weighted means with weights that weaken the contribution of those who tend to be far from the normal mean and amplify the contribution of those who tend to be closer to the normal mean.
}
\examples{
x <- rnorm(1000,0,1)
hypermean(x,size=0.1,type="trim")
}
