% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypermeanr-package.R
\docType{package}
\name{hypermeanr-package}
\alias{hypermeanr}
\alias{hypermeanr-package}
\title{hypermeanr: Function to calculate hypermean}
\description{
Provides a function to calculate hyperman, weighted means with weights that weaken the contribution of those who tend to be far from the normal mean and amplify the contribution of those who tend to be closer to the normal mean.
}
\author{
\strong{Maintainer}: Keita Abe \email{keita43a@gmail.com} (\href{https://orcid.org/0000-0002-1833-3036}{ORCID})

}
\keyword{internal}
